ci(clippy): enforce warning-free workspace lint gate

+++ onchain/contracts/multisig/src/lib.rs
@@ -488,7 +488,7 @@ impl MultisigContract {
             OperationType::DisputeResolution,
         ] {
             if let Some(override_val) = read_threshold_override(&env, &op_type) {
-                if override_val > signer_count as u32 {
+                if override_val > signer_count {
                     env.storage()
                         .persistent()
                         .set(&StorageKey::ThresholdOverride(op_type), &signer_count);
